d_a_npc_aj1: bodyCreateHeap 98.92->99.22 NORM-DIFF-0 — (1) EMode_NULL not EMode_LOOP in McaMorf ctor (retail r9=-1; mwcc ctor-arg scheduling puts loopMode in arg7 slot, hi1 precedent); (2) init_texPttrnAnm return type s32->u8 (retail callers clrlwi.24 the result; fn tail stays subic/subfe no mask); fresh .o postdates

+++ include/d/actor/d_a_npc_aj1.h
@@ -35,7 +35,7 @@ public:
     void setMtx(bool);
     int bckResID(int);
     int btpResID(int);
-    s32 init_texPttrnAnm(signed char, bool);
+    u8 init_texPttrnAnm(signed char, bool);
     void play_texPttrnAnm();
     void setAnm_anm(anm_prm_c*);
     void setAnm_NUM(int, int);
